refactor: consolidate bulk utils
