+++ messages/messages.md
@@ -1,3 +1,28 @@
+# success
+
+Bulk %s request %s started successfully
+
+# checkStatus
+
+Run command sf data %s resume -i %s -o %s to check status.
+
+# checkJobViaUi
+
+To review the details of this job, run:
+sf org open --target-org %s --path "/lightning/setup/AsyncApiJobStatus/page?address=%2F%s"
+
+# remainingTimeStatus
+
+Remaining time: %d minutes
+
+# remainingRecordsStatus
+
+Processed %d | Success %d | Fail %d
+
+# bulkJobFailed
+
+The bulk job %s failed. Check the job status for more information.
+
 # perfLogLevelOption
 
 Get API performance data.
